Moved filterButton to Tournament List dir
