+++ src/MainScreen/Tournament Search/FilterButton.tsx
@@ -6,7 +6,7 @@ import Animated, { useAnimatedStyle, useSharedValue, withSpring } from "react-na
 import { TransparentCard } from "../../Shared"
 import { IoniconsThemed } from "../../Shared/IconTheme"
 import { MainText } from "../../Shared/Text"
-import { FilterButtonProps, FilterButtonRefProps } from "./types"
+import { FilterButtonProps, FilterButtonRefProps } from "../Debug/types"
 
 const MAX_WIDTH = 200;
